PremiumWatch/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
coverage==7.6.1
crispy-bootstrap5==2024.2
cycler==0.12.1
Django==4.2.15
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
Faker==27.0.0
fonttools==4.53.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.2
iniconfig==2.0.0
kiwisolver==1.4.5
Markdown==3.7
matplotlib==3.9.2
mock==5.1.0
networkx==3.2.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zipp==3.20.0
